Crate payment_settle_accounts

Re-exports

pub use crate::program::*;

Modules

program

This module contains the definition of the pipeline trait and its implementations for running a program that reads transactions from some Source, process them with some PaymentEngine, and writes to some Sink.

Structs

CSVTransactionReader

CSVTransactionReader is a wrapper around csv::Reader.

CSVTransactionResultStdoutWriter

CSVTransactionResultStdoutWriter is a wrapper around csv::Writer using stdout.

CentDenomination

Represents a monetary value in cents denomination to avoid floating point precissions issues.

MemoryThreadSafePaymentEngine

A thread-safe payment engine that stores transaction information in memory. State is protected by a RwLock to allow concurrent reads and exclusive writes in order to speed up the processing of transactions.

Transaction

Represents a transaction object.

TransactionResult

Represents the result of a transaction.

Enums

TransactionError

Error type for the transaction processing based on thiserror crate

TransactionType

Represents the type of a transaction.

Traits

PaymentEngine

Trait representing a payment engine. PaymentEngine is responsible for processing transactions one by one and keeping track of them in a TransactionResult per Client Account.

Sink

Source

Type Definitions

ClientId

Represents a client ID.

TxId

Represents a transaction ID.